ar/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astunparse==1.6.3
base-url==0.0.1
beautifulsoup4==4.9.3
bleach==3.3.0
build==0.3.1.post1
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
cython==3.0a6
distlib==0.3.1
docutils==0.17
filelock==3.0.12
gast==0.3.3
glcontext==2.3
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
icrawler==0.6.3
idna==2.9
imageio==2.9.0
importlib-metadata==3.10.0
imutils==0.5.3
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lxml==4.6.1
Markdown==3.2.2
matplotlib==3.3.3
moderngl==5.6.3
MouseInfo==0.1.3
mysql-connector-python==8.0.22
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.9
pandas==1.2.2
pep517==0.10.0
Pillow==8.1.0
pkginfo==1.7.0
playsound==1.2.2
prettytable==2.0.0
progressbar==2.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycryptodome==3.9.9
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.8.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scipy==1.4.1
SimpleWebSocketServer @ git+https://github.com/dpallot/simple-websocket-server.git@a6ddef766a4f5fa655c9072107fccc28ecf2b3e3
six==1.15.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape@8c2c0fa47a053c351f120e98dbea3a4eadb7035b
soupsieve==2.0.1
streamlink==2.0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tf-slim==1.1.0
toml==0.10.2
tqdm==4.60.0
twine==3.4.1
urllib3==1.25.9
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
win10toast==0.9
wrapt==1.12.1
zipp==3.4.1
